Reproduced the some of the behaviour observed in the browser. Structural changes to allow easier header adjustments
I've identified the state syncronization mechanism if not its inner
workings. The browser periodically (~10 seconds) sends a post request
to the server with form data with variable __VIEWSTATE equal to some
encoded information. The wiki article suggest this is GUI state,
which is reasonable considering the request is immediatly sent after
a GUI effect.

Its getting to be 4am, I think I've obsessed over this problem enough
today.