Populates Database with scheduled stops Saves cookie in db while valid
After fighting with a bug, I discovered a weird phenomenon.
I assumed that any referenced route in scheduledStops
must exist, this is not the case. There is at least one
'ghost route', #51 that appears in the schedule for various
stops but does not seem to exist in any records on the pvta
site or schedules. It seems that this bus passes through
North Amherst, I may visit it just to see if it exists and
the name of the company that runs it.